+++ Services/Ordering.Application/Features/Commands/UpdateOrder/UpdateOrderCommand.cs
@@ -6,9 +6,9 @@ public class UpdateOrderCommand : IRequest
 {
     public int Id { get; set; }
     public string UserName { get; set; } = default!;
-    public decimal TotalPrice { get; set; }
+    public double TotalPrice { get; set; }
 
-    // BillingAddress 
+    // BillingAddress
     public string FirstName { get; set; } = default!;
     public string LastName { get; set; } = default!;
     public string EmailAddress { get; set; } = default!;
